refactor(gateway): isolate credential authorizer
Move credential authorization lifecycle and request scoping behind a topical mixin while keeping the API server as the composition facade. Split the oversized credential contract suite by behavior without dropping any cases.
